% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegYan-package.R
\docType{package}
\name{ggsegYan-package}
\alias{ggsegYan}
\alias{ggsegYan-package}
\title{ggsegYan: Yan 2023 Homotopic Atlas for the 'ggseg' Ecosystem}
\description{
Yan 2023 homotopic brain parcellation for the 'ggseg' ecosystem. Provides 'ggseg_atlas' objects with 2D polygon geometry and 3D vertex indices. Includes 7-network and 17-network variants at resolutions from 100 to 1000 parcels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegYan}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegYan/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
