% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{yan7_800}
\alias{yan7_800}
\title{Yan 7-Network 800-Parcel Atlas}
\usage{
yan7_800()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Yan et al. (2023) homotopic cortical parcellation
with 800 parcels aligned to 7 resting-state networks (Yeo et al., 2011).
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
yan7_800()
plot(yan7_800())
}
\references{
Yan X, et al. (2023). "Homotopic local-global parcellation
of the human cerebral cortex from resting-state functional connectivity."
\emph{NeuroImage}, 273:120010.
\doi{10.1016/j.neuroimage.2023.120010}
}
\seealso{
Other ggseg_atlases: 
\code{\link{yan17_100}()},
\code{\link{yan17_1000}()},
\code{\link{yan17_200}()},
\code{\link{yan17_300}()},
\code{\link{yan17_400}()},
\code{\link{yan17_500}()},
\code{\link{yan17_600}()},
\code{\link{yan17_700}()},
\code{\link{yan17_800}()},
\code{\link{yan17_900}()},
\code{\link{yan7_100}()},
\code{\link{yan7_1000}()},
\code{\link{yan7_200}()},
\code{\link{yan7_300}()},
\code{\link{yan7_400}()},
\code{\link{yan7_500}()},
\code{\link{yan7_600}()},
\code{\link{yan7_700}()},
\code{\link{yan7_900}()}

Other cortical_atlases: 
\code{\link{yan17_100}()},
\code{\link{yan17_1000}()},
\code{\link{yan17_200}()},
\code{\link{yan17_300}()},
\code{\link{yan17_400}()},
\code{\link{yan17_500}()},
\code{\link{yan17_600}()},
\code{\link{yan17_700}()},
\code{\link{yan17_800}()},
\code{\link{yan17_900}()},
\code{\link{yan7_100}()},
\code{\link{yan7_1000}()},
\code{\link{yan7_200}()},
\code{\link{yan7_300}()},
\code{\link{yan7_400}()},
\code{\link{yan7_500}()},
\code{\link{yan7_600}()},
\code{\link{yan7_700}()},
\code{\link{yan7_900}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
